requirements.txt
numpy
scipy
ffmpeg-python
openai-whisper
TTS
pydub
